Configured SQL server and made the relationship between Ticker and Company one-to-one

+++ StockExchangeWebsite/Models/StockMarket.cs
@@ -13,10 +13,10 @@ namespace StockExchangeWebsite.Models
         public TimeSpan OpeningTime { get; set; }
         public TimeSpan ClosingTime { get; set; }
         public List<DayOfWeek> TradingDays { get; set; }
-        public decimal Liquidity { get; set; }
-        public decimal TransactionFee { get; set; }
-        public decimal TickSize { get; set; }
-        public decimal MarketCapitalization { get; set; }
+        public double Liquidity { get; set; }
+        public double TransactionFee { get; set; }
+        public double TickSize { get; set; }
+        public double MarketCapitalization { get; set; }
         public List<OrderType> SupportedOrderTypes { get; set; }
         public int SettlementPeriod { get; set; }
         public float Volatility { get; set; }
@@ -26,7 +26,7 @@ namespace StockExchangeWebsite.Models
         public int MarketDepth { get; set; }
         public bool AllowsShortSelling { get; set; }
         public bool HasCircuitBreakers { get; set; }
-        public decimal CircuitBreakerThreshold { get; set; }
+        public double CircuitBreakerThreshold { get; set; }
 
         //Relations
         public List<StockMarketTicker> StockMarketTickers { get; set; }
